+++ app/inc/AppConfig.hpp
@@ -12,6 +12,7 @@
 #include <any>
 #include <algorithm>
 #include <iterator>
+#include <map>
 #include <type_traits> // for std::remove_pointer
 #include <glog/logging.h>
 
@@ -21,8 +22,11 @@ struct AppConfig {
     std::string labelsPath;
     std::string tokenizerVocabPath;
     std::string tokenizerMergesPath;
+    std::string bertTokenizerVocabPath;
     std::string weights;
+    std::string mmprojectPath;
     std::vector<std::string> textPrompts;
+    std::map<std::string, std::string> taskExtraParams;
     bool use_gpu;
     bool enable_warmup;
     bool enable_benchmark;